Add a simple HTML view of the healths
Using freemarker templates and a tiny bit of bootstrap to make it a tiny
bit nicer than just text.

Tried out mustache templates which where quite nice, but the missing
ifs make it really hard to do something useful unless you put a lot of
more intermediate representations in the view model.

Refactored the API so that healths now is a list of values which all
have a name, replacing the key/value pairs with name/health.

Close #2 for now...