Move logic around state state of system, services and nodes to domain
This includes
 * Keeping track of which services and nodes are deployed
 * What's the state

Also separate state transactions from events, and make the whole system
kind of event sourced, although it's not really events that are stared
and the (non) events are pruned.

Switched to synchronous EventBus to make everything a bit easier to
reason about. If external comminication will start to happen based on
events, the events could be propagated to an asynch bus for that
purpose.

Suddenly, a great need of separate the logic from view model arrised...
And thinking about it, will still kind of only have a view model but now
we call it the domain, and we expose the domain straight out to the
public (limited parts of it at least). To make this really clean, we
need #13, and more descriptive events about the whole state of the
system which will make it easier on the clients (the Resource, or
notifiers to figure out how to update their state).

Part of #12