Build a simple notification system
... where external parties can take part of health changed events.

It's a simple webhook system, wheer you register the root endpoint of
your hook and the respective event endpoints will be called as events
happen in the system.

Next step is to decouple the synchronous internal event bus from this
notification system, so that unreliable hooks don't hog and break this
service.

First part of #4